Add indexed UTxO type.
The index makes it possible to efficiently compute the subset of a UTxO
set containing a particular asset, or to select a UTxO entry containing
that asset, without having to search through the entire UTxO set.

The index also keeps track of the current UTxO balance for all assets,
making it possible to efficiently look up the quantity of a particular
asset without having to sum across the entire UTxO set.