Simplify representation of inputs in `SelectionSkeleton`.
The purpose of `SelectionSkeleton` is to provide all the data necessary
to calculate the fee associated with a selection.

However, this calculation only depends on the number of inputs, and not
the contents of each input.

Therefore, it's only necessary to include a count of the selected inputs
within `SelectionSkeleton`, and unnecessary to include the actual inputs
themselves.