Minimize the responsibility of `makeChangeForNonUserSpecifiedOutputs`.
This change moves the following responsibilities out of the
`makeChangeForNonUserSpecifiedOutputs` function:

    - the removing of burned values
    - the addition of minted values

This reduces the complexity of property and unit tests that test this
function.

The following functions are still checked by relevant properties:

    - `addMintValuesToChangeMaps`
    - `removeBurnValuesFromChangeMaps`

Further changes will add unit tests to act as documentation and sanity
checks for the above functions.