Move minted/burned assets from `SelectionSkeleton` to `TransactionCtx`.
The `estimateTxSize` function is provided with a `TxSkeleton`, which is
constructed from a `SelectionSkeleton` and a `TransactionCtx`.

Therefore, provided that a `TxSkeleton` is constructed with the correct
maps of minted and burned assets, the `estimateTxSize` function does not
mind whether these values originate from a `SelectionSkeleton` or from a
`TransactionCtx`.

However, there are strong advantages to storing minted and burned assets
within `TransactionCtx` instead of `SelectionSkeleton`:

* Minted and burned assets remain constant over the course of an ongoing
  coin selection.

* Therefore, the size and cost of minted and burned assets is constant
  over the course of an ongoing coin selection.

* The `{gen,shrink}SelectionSkeleton` functions will no longer need to
  generate and shrink minted and burned assets.

* The mock `computeMinimumCostLinear` function will no longer need to
  determine the sizes of minted and burned assets, leading to faster
  execution times for `BalanceSpec`.

* The `selectAssets` function will be able to populate the `assetsToMint`
  and `assetsToBurn` fields of the `SelectionParams` object, since it
  already has access to a `TransactionCtx` value.