Fork the `SelectionParams` type.
This commit forks the `SelectionParams` into two variants:

1. A wallet-specific variant, with wallet-friendly types, carrying
   context that is important to the creation of a wallet-friendly
   final `Selection` value.

2. An internal variant that will eventually not depend on wallet types
   and not carry context that is unimportant to coin selection.

This commit also provides a function `toInternalSelectionParams` to
convert from the wallet-friendly type to the internal type.

Currently, this function is a no-op, but it will eventually strip
unnecessary context away so that the coin selection library does not
have to deal with it.