Change type of `Selection.inputs` to `NonEmpty (TxIn, TokenBundle)`.
The type of `Selection.inputs` is currently `NonEmpty (TxIn, TxOut)`.

However the coin selection library does not need to manipulate the
addresses associated with these inputs.

The caller of coin selection already has all the relevant context
required to resolve the addresses of selected inputs, since it has
access to the available UTxO set.

Therefore, we can remove addresses from the list of selected inputs
returned.

This takes us a step closer to removing the coin selection library's
dependence on the `TxOut` and `Address` types.