Rewrite balanceTransaction using evalauteTransactionBalance
The core idea is to 1) evaluate the balance of the partial tx using the
already existing `evaluateTransactionBalance` and then 2) ask
coin-selection to select enough inputs to cover it.

The key pieces of code are:

```
(balance0, fee0) <- balanceAfterSettingMinFee partialTx
```

```
    { assetsToMint = positiveTokens <> tokensInOutputs
    , assetsToBurn = negativeTokens <> tokensInInputs
    , extraCoinIn =
        positiveAda
        <> adaInOutputs
        <> fromCardanoLovelace fee0
    , extraCoinOut = negativeAda <> adaInInputs
    }
```

The effects are:
1. Can now balance transactions containing deposits and reclaims (e.g.
  first time delegation txs)
2. Enabling of the `prop_balanceTransactionBalanced` property
3. Fixes of several variable-length coin-encoding related corner cases,
   which previously would have resulted in internal server errors or
   improperly balanced results.

Note that the maximum transaction size is not respsected in this commit.
Future commits will fix that.