Rename function `utxoFromUnvalidatedTx` to `utxoFromTxOutputs`.
This name still reflects the purpose, namely that it generates a UTxO
set from the ordinary outputs of a transaction (without regard for its
validation status).

However, the new name allows us to create an obviously-related sister
function called `utxoFromTxCollateralOutputs`.

This sister function will be introduced in a later commit.