Only give `prop_changeUTxO_inner` txs that have not failed script validation.
The function `prop_changeUTxO_inner` expects a list of **pending**
transactions.

Pending transactions, by definition, are those that have not yet been
confirmed as having been included on the blockchain, and so therefore we
cannot know their script validation status.

However, because `genTx` can generate transactions that are marked as
having failed script validation, we cannot use this generator directly
if we **only** want pending transactions. Instead, we must modify the
generator so that it only generates transactions with a `scriptValidity`
field equal to `Nothing`.

(It would be better if our transaction types made it impossible for a
pending transaction to be marked as having failed script validation, but
that is arguably a change for a future commit.)